Best separation of unit and functional tests; better cleanup
